refactor unsafe ga_fence stuff into separate source file; reconf
