+++ global/src/onesided.c
@@ -81,11 +81,6 @@
 /*uncomment line below to verify consistency of MA in every sync */
 /*#define CHECK_MA yes */
 
-#ifndef DISABLE_UNSAFE_GA_FENCE
-char *fence_array; /* RACE */
-static int GA_fence_set=0; /* RACE */
-#endif // DISABLE_UNSAFE_GA_FENCE
-
 extern void armci_read_strided(void*, int, int*, int*, char*);
 extern void armci_write_strided(void*, int, int*, int*, char*);
 extern armci_hdl_t* get_armci_nbhandle(Integer *);
@@ -109,25 +104,15 @@ void pnga_pgroup_sync(Integer grp_id)
           ARMCI_Fence(ARMCI_Absolute_id(&PGRP_LIST[grp_id].group, p));
        }
        pnga_msg_pgroup_sync(grp_id);
-#  ifndef DISABLE_UNSAFE_GA_FENCE
-       if (GA_fence_set) {
-           bzero(fence_array,(int)GAnproc);
-       }
-       GA_fence_set=0;
-#  endif // DISABLE_UNSAFE_GA_FENCE
 #else // MSG_COMMS_MPI
+       gai_fence_reset();
        pnga_error("ga_pgroup_sync_(): MPI not defined. ga_msg_pgroup_sync_()  can be called only if GA is built with MPI", 0);
 #endif // MSG_COMMS_MPI
     } else {
        /* printf("p[%d] calling regular sync in ga_pgroup_sync\n",GAme); */
        ARMCI_AllFence();
        pnga_msg_pgroup_sync(grp_id);
-#  ifndef DISABLE_UNSAFE_GA_FENCE
-       if (GA_fence_set) {
-           bzero(fence_array,(int)GAnproc);
-       }
-       GA_fence_set=0;
-#  endif // DISABLE_UNSAFE_GA_FENCE
+       gai_fence_reset();
     }
 #ifdef CHECK_MA
     Integer status = MA_verify_allocator_stuff();
@@ -146,12 +131,7 @@ void pnga_sync(void)
     if (GA_Default_Proc_Group == -1) {
       ARMCI_AllFence();
 	  pnga_msg_sync();
-#  ifndef DISABLE_UNSAFE_GA_FENCE
-	  if(GA_fence_set) {
-          bzero(fence_array,(int)GAnproc);
-      }
-	  GA_fence_set=0;
-#  endif // DISABLE_UNSAFE_GA_FENCE
+      gai_fence_reset();
     } else {
           Integer grp_id = (Integer)GA_Default_Proc_Group;
           pnga_pgroup_sync(grp_id);
@@ -160,62 +140,6 @@ void pnga_sync(void)
     Integer status = MA_verify_allocator_stuff();
 #endif
 }
-
-#ifndef DISABLE_UNSAFE_GA_FENCE
-
-/**
- *  Wait until all requests initiated by calling process are completed
- */
-#if HAVE_SYS_WEAK_ALIAS_PRAGMA
-#   pragma weak wnga_fence = pnga_fence
-#endif
-
-void pnga_fence(void)
-{
-    if(GA_fence_set<1) {
-        pnga_error("ga_fence: fence not initialized",0);
-    }
-
-    /* Why does this not set the fence array to zero? */
-    GA_fence_set--;
-
-    for(int proc=0; proc<GAnproc; proc++) {
-        if(fence_array[proc]) {
-            ARMCI_Fence(proc);
-        }
-    }
-    bzero(fence_array,(int)GAnproc);
-    return;
-}
-
-/**
- *  Initialize tracing of request completion
- */
-#if HAVE_SYS_WEAK_ALIAS_PRAGMA
-#   pragma weak wnga_init_fence = pnga_init_fence
-#endif
-
-void pnga_init_fence(void)
-{
-    /* Why is this not setting it to 1? */
-    GA_fence_set++;
-}
-
-void gai_init_onesided(void)
-{
-    fence_array = calloc((size_t)GAnproc,1);
-    if(!fence_array)
-        pnga_error("ga_init:calloc failed",0);
-}
-
-
-void gai_finalize_onesided(void)
-{
-    free(fence_array);
-    fence_array = NULL;
-}
-
-#endif // DISABLE_UNSAFE_GA_FENCE
 
 /*\ prepare permuted list of processes for remote ops
 \*/
@@ -653,9 +577,7 @@ void ngai_put_common(Integer g_a,
             /* BJP */
             proc = PGRP_LIST[p_handle].inv_map_proc_list[proc];
           }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-          if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+          gai_fence_set(proc);
 
 #ifdef PERMUTE_PIDS
           if(GA_Proc_list) proc = GA_inv_Proc_list[proc];
@@ -787,9 +709,7 @@ void ngai_put_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -971,9 +891,7 @@ void ngai_put_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -1508,9 +1426,7 @@ void ngai_get_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -1691,9 +1607,7 @@ void ngai_get_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -1938,9 +1852,7 @@ void ngai_acc_common(Integer g_a,
             proc = (int)PGRP_LIST[p_handle].inv_map_proc_list[proc];
           }
 
-# ifndef DISABLE_UNSAFE_GA_FENCE
-          if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+          gai_fence_set(proc);
 
 #ifdef PERMUTE_PIDS
           if(GA_Proc_list) proc = GA_inv_Proc_list[proc];
@@ -2061,9 +1973,7 @@ void ngai_acc_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -2234,9 +2144,7 @@ void ngai_acc_common(Integer g_a,
                 gam_setstride(ndim, size, ld, ldrem, stride_rem, stride_loc);
 
                 proc = pinv;
-# ifndef DISABLE_UNSAFE_GA_FENCE
-                if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+                gai_fence_set(proc);
 
 #ifndef NO_GA_STATS	    
                 if(proc == GAme){
@@ -3120,9 +3028,7 @@ void gai_scatter_acc_local(Integer g_a, void *v,Integer *i,Integer *j,
   desc.dst_ptr_array = ptr_dst;
   desc.ptr_array_len = (int)nv;
 
-# ifndef DISABLE_UNSAFE_GA_FENCE
-  if(GA_fence_set)fence_array[proc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+  gai_fence_set(proc);
 
 #ifdef PERMUTE_PIDS
     if(GA_Proc_list) proc = GA_inv_Proc_list[proc];
@@ -3808,9 +3714,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
             } else {
               iproc = PGRP_LIST[p_handle].inv_map_proc_list[aproc[k]];
             }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-            if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+            gai_fence_set(iproc);
 
             rc=ARMCI_PutV(&desc, 1, (int)iproc);
             if(rc) pnga_error("scatter failed in armci",rc);
@@ -3830,9 +3734,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
               if (p_handle >= 0) {
                 iproc = PGRP_LIST[p_handle].inv_map_proc_list[iproc];
               }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-              if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+              gai_fence_set(iproc);
 
               rc=ARMCI_PutV(&desc, 1, (int)iproc);
               if(rc) pnga_error("scatter failed in armci",rc);
@@ -3856,9 +3758,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
               if (p_handle >= 0) {
                 iproc = PGRP_LIST[p_handle].inv_map_proc_list[iproc];
               }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-              if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+              gai_fence_set(iproc);
 
               rc=ARMCI_PutV(&desc, 1, (int)iproc);
               if(rc) pnga_error("scatter failed in armci",rc);
@@ -3929,9 +3829,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
             } else {
               iproc = PGRP_LIST[p_handle].inv_map_proc_list[aproc[k]];
             }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-            if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+            gai_fence_set(iproc);
 
             if(alpha != NULL) {
               int optype=-1;
@@ -3961,9 +3859,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
               if (p_handle >= 0) {
                 iproc = PGRP_LIST[p_handle].inv_map_proc_list[iproc];
               }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-              if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+              gai_fence_set(iproc);
 
               if(alpha != NULL) {
                 int optype=-1;
@@ -3997,9 +3893,7 @@ void gai_gatscat(int op, Integer g_a, void* v, Integer subscript[],
               if (p_handle >= 0) {
                 iproc = PGRP_LIST[p_handle].inv_map_proc_list[iproc];
               }
-# ifndef DISABLE_UNSAFE_GA_FENCE
-              if(GA_fence_set) fence_array[iproc]=1;
-# endif // DISABLE_UNSAFE_GA_FENCE
+              gai_fence_set(iproc);
 
               if(alpha != NULL) {
                 int optype=-1;
